Migrate AWS and EKS providers to AWS SDK v2 for Go (#10922)
* Initial migration to AWS SDK v2


* Port authenticator code to aws-sdk-go-v2


* Fix last compile issues


* Update codegen


* Fix import statement ordering


* Fix linting issues


* Prevent nil pointer panic


* Fix missed lint issue


* Make permission formatting npe-proof


* Next attempt at fixing integration tests


* Write a custom compare function for ip permissions


* Fix import statement


* post-rebase blues

* remove error decoding

smithy.APIError implementations all seem to be doing the same, if not
better string marshalling that we would do.

Also using HTTP 500 (internal server error) for every possible fault
is not ideal (ValidateCredentials() should not error out with an internal
server error on bad credentials), and in general the code in the
provider package should not return HTTP-related things, as controllers
also use it.

I removed the DecodeError() entirely, as it seems to serve no purpose
anymore with SDK v2. :)

* Only read from array if length is 2 and set the correct option


* Revert "remove error decoding"

This reverts commit 628c8727b937b7eb338b1eb009283f3ca61d56da.

* Rework DecodeError to include HTTP status codes


* Make linter happy

+++ pkg/test/e2e/utils/apiclient/models/e_k_s_nodegroup_scaling_config.go
@@ -18,17 +18,17 @@ import (
 type EKSNodegroupScalingConfig struct {
 
 	// The current number of nodes that the managed node group should maintain.
-	DesiredSize int64 `json:"desiredSize,omitempty"`
+	DesiredSize int32 `json:"desiredSize,omitempty"`
 
 	// The maximum number of nodes that the managed node group can scale out to.
 	// For information about the maximum number that you can specify, see Amazon
 	// EKS service quotas (https://docs.aws.amazon.com/eks/latest/userguide/service-quotas.html)
 	// in the Amazon EKS User Guide.
-	MaxSize int64 `json:"maxSize,omitempty"`
+	MaxSize int32 `json:"maxSize,omitempty"`
 
 	// The minimum number of nodes that the managed node group can scale in to.
 	// This number must be greater than zero.
-	MinSize int64 `json:"minSize,omitempty"`
+	MinSize int32 `json:"minSize,omitempty"`
 }
 
 // Validate validates this e k s nodegroup scaling config
